Taking rounding errors into account. Output of Agro model agrees with output from tool wrapper simulation within rounding errors as shown by diff.

+++ core/fmu-import-export/src/main/resources/c-templates/Fmu.h
@@ -27,9 +27,9 @@ struct FmiBuffer{
 
 struct PeriodicThreadStatus
 {
-	double period;
+	long double period;
 	void (*call)(void);
-	double lastExecuted;
+	long double lastExecuted;
 };
 
 extern struct FmiBuffer fmiBuffer;
